fix: use correct type for openh264enc gop-size property
openh264enc expects guint (u32) for gop-size, not gint (i32).
This caused a panic on pipeline startup.

+++ src/gstreamer/pipeline.rs
@@ -338,7 +338,7 @@ impl VideoPipeline {
             if encoder.has_property("key-int-max", None) {
                 let _ = encoder.set_property("key-int-max", interval as i32);
             } else if encoder.has_property("gop-size", None) {
-                let _ = encoder.set_property("gop-size", interval as i32);
+                let _ = encoder.set_property("gop-size", interval);
             } else if encoder.has_property("keyframe-max-dist", None) {
                 let _ = encoder.set_property("keyframe-max-dist", interval as i32);
             } else if encoder.has_property("keyframe-period", None) {
